test(react-query): use vitest typecheck correctly with *.test-d.ts (#6978)
