+++ package.json
@@ -50,7 +50,7 @@
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^6.20.0",
     "@typescript-eslint/parser": "^6.20.0",
-    "@vitest/coverage-istanbul": "^1.2.2",
+    "@vitest/coverage-istanbul": "^1.3.1",
     "cpy-cli": "^5.0.0",
     "esbuild-plugin-file-path-extensions": "^2.0.0",
     "eslint": "^8.56.0",
@@ -73,7 +73,7 @@
     "tsup": "^8.0.1",
     "typescript": "5.2.2",
     "vite": "^5.1.1",
-    "vitest": "^1.2.2"
+    "vitest": "^1.3.1"
   },
   "pnpm": {
     "overrides": {
